feat(users): finalize user table implementation across all layers

+++ ROSCA.Application/DTOs/Users/IntegrityProfileDTO.cs
@@ -14,7 +14,7 @@ namespace ROSCA.Application.DTOs.Users
         public int LatePaymentsCount { get; set; }
         public int MissingPayments { get; set; }
         public double CommitmentRate { get; set; }
-        public Decimal RawScore { get; set; }
+        public double  RawScore { get; set; }
         public string Level { get; set; } = string.Empty;
     }
 }
